@@ -45,13 +45,13 @@
     "release:major": "standard-version -r major",
     "release:minor": "standard-version -r minor",
     "release:patch": "standard-version -r patch",
-    "release:alpha": "standard-version --prerelease alpha"
+    "release:beta": "standard-version --prerelease beta"
   },
   "resolutions": {
     "react-error-overlay": "6.0.9"
   },
   "peerDependencies": {
-    "react": "17.0.2"
+    "react": "18.1.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^4.2.4",
@@ -59,8 +59,8 @@
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^25.1.4",
     "@types/node": "15.0.3",
-    "@types/react": "17.0.5",
-    "@types/react-dom": "^16.9.7",
+    "@types/react": "18.0.9",
+    "@types/react-dom": "18.0.4",
     "@types/react-transition-group": "^4.4.4",
     "babel-eslint": "^10.0.3",
     "bear-jsutils": "1.0.1",
@@ -71,9 +71,9 @@
     "microbundle-crl": "^0.13.10",
     "npm-run-all": "^4.1.5",
     "classnames": "^2.3.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "4.0.3",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.2",
     "standard-version": "9.3.0",
     "typescript": "4.1.5"
